+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - Added cluster id to emr step metrics
 - Added two handlers for increasing/decreasing desired capacity of autoscaling groups
 - Implemented check for reserved instances
+- Added check to ensure that some or all AWS ConfigService rules have full compliance
 
 ## [2.3.0] - 2016-03-18
 ### Added